requirements.txt
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
Pillow==10.2.0
streamlit==1.30.0
weaviate_client
